Move GsonConfig to util package
