+++ java/clients/src/main/java/sleeper/status/report/CompactionTaskStatusReport.java
@@ -19,13 +19,13 @@ import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import sleeper.ClientUtils;
 import sleeper.compaction.status.store.task.DynamoDBCompactionTaskStatusStore;
 import sleeper.compaction.task.CompactionTaskStatusStore;
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.status.report.compaction.task.CompactionTaskQuery;
 import sleeper.status.report.compaction.task.CompactionTaskStatusReportArguments;
 import sleeper.status.report.compaction.task.CompactionTaskStatusReporter;
+import sleeper.util.ClientUtils;
 
 import java.io.IOException;
