+++ java/clients/src/main/java/sleeper/status/report/FilesStatusReport.java
@@ -20,7 +20,6 @@ import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import org.apache.hadoop.conf.Configuration;
-import sleeper.ClientUtils;
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.configuration.properties.table.TablePropertiesProvider;
 import sleeper.statestore.StateStore;
@@ -32,6 +31,7 @@ import sleeper.status.report.filestatus.FileStatusCollector;
 import sleeper.status.report.filestatus.FileStatusReporter;
 import sleeper.status.report.filestatus.JsonFileStatusReporter;
 import sleeper.status.report.filestatus.StandardFileStatusReporter;
+import sleeper.util.ClientUtils;
 
 import java.io.IOException;
 import java.util.HashMap;
