+++ java/clients/src/main/java/sleeper/status/report/IngestTaskStatusReport.java
@@ -19,13 +19,13 @@ import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import sleeper.ClientUtils;
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.ingest.status.store.task.DynamoDBIngestTaskStatusStore;
 import sleeper.ingest.task.IngestTaskStatusStore;
 import sleeper.status.report.ingest.task.IngestTaskQuery;
 import sleeper.status.report.ingest.task.IngestTaskStatusReportArguments;
 import sleeper.status.report.ingest.task.IngestTaskStatusReporter;
+import sleeper.util.ClientUtils;
 
 import java.io.IOException;
