+++ java/clients/src/main/java/sleeper/status/report/TrackedQueryProgressReport.java
@@ -19,11 +19,11 @@ import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import sleeper.ClientUtils;
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.query.tracker.DynamoDBQueryTracker;
 import sleeper.query.tracker.exception.QueryTrackerException;
 import sleeper.statestore.StateStoreException;
+import sleeper.util.ClientUtils;
 
 import java.io.IOException;
