+++ java/clients/src/main/java/sleeper/status/report/compaction/job/CompactionJobStatusReportArguments.java
@@ -26,7 +26,7 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
 
-import static sleeper.ClientUtils.optionalArgument;
+import static sleeper.util.ClientUtils.optionalArgument;
 
 public class CompactionJobStatusReportArguments {
     private static final String DEFAULT_REPORTER = "STANDARD";
