+++ java/clients/src/main/java/sleeper/util/GsonConfig.java
@@ -13,12 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.status.report.job;
+package sleeper.util;
 
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonPrimitive;
 import com.google.gson.JsonSerializer;
-import sleeper.core.record.process.RecordsProcessedSummary;
 
 import java.time.Duration;
 import java.time.Instant;
@@ -32,7 +31,6 @@ public class GsonConfig {
         return new GsonBuilder().serializeSpecialFloatingPointValues()
                 .registerTypeAdapter(Instant.class, instantJsonSerializer())
                 .registerTypeAdapter(Duration.class, durationSerializer())
-                .registerTypeAdapter(RecordsProcessedSummary.class, JsonRecordsProcessedSummary.serializer())
                 .setPrettyPrinting();
     }
